// enumeratewebapp/src/model/EnumerateSK2.ts
export interface EnumerateSK2 {
    AREA_CODE: string;
    REG: string;
    REGN: string;
    CWT: string;
    CWTN: string;
    AMP: string;
    AMPN: string;
    TAM: string;
    TAMN: string;
    TYPE: string;
    MUN: string;
    MUNN: string;
    VIL: string;
    VILN: string;
    VILT: string;
    ADD: string;
    AH_CODE: string;
    HTIT: string;
    HTIT_T: string;
    HNAMEF: string;
    HNAMES: string;
    AH_NO: string;
    REF: string;
    TEL: string;
    email: string;
    HOL: string;
    HTIT_A: string;
    HTIT_TA: string;
    HNAMEF_A: string;
    HNAMES_A: string;
    TEL_A: string;
    email_A: string;
    SDATE_A: string;
    SDATE_B: string;
    SDATE_C: string;
    STIME_A: string;
    STIME_B: string;
    STIME_C: string;
    CDATE_A: string;
    CDATE_B: string;
    CDATE_C: string;
    CTIME_A: string;
    CTIME_B: string;
    CTIME_C: string;
    P1_A: string;
    P1_B: string;
    P1_C: string;
    P2_A: string;
    P2_B: string;
    P2_C: string;
    P3_A: string;
    P3_B: string;
    P3_C: string;
}